requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cachetools==5.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
decorator==4.4.2
distro==1.9.0
feedparser==6.0.11
google-api-core==2.24.2
google-api-python-client==2.108.0
google-auth==2.39.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.70.0
h11==0.14.0
httpcore>=0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.6
imageio==2.31.2
imageio-ffmpeg==0.5.1
joblib==1.3.2
moviepy==1.0.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.76.0
pillow==9.5.0
proglog==0.1.11
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pyparsing==3.2.3
python-dotenv==1.0.0
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.7.2
sgmllib3k==1.0.0
sniffio==1.3.1
tqdm==4.66.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.2.1
google-cloud-texttospeech==2.16.2
google-generativeai==0.8.5
